Close small mutation gaps in the script helpers, and split the summary tests
Adds the missing checks for the shared start-up files, worker counts, the
project root, stream decoding, and the file lock's exclusivity. The mutation
summary's tests now sit in three files, one per report.
